pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rna3d-local"
version = "0.1.0"
description = "RNA 3D local tooling (phase 1 oracle retrieval)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "polars>=1.8",
  "pyarrow>=18.0",
  "torch>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
llm = [
  "llama-cpp-python>=0.2.90",
]
faiss = [
  "faiss-cpu>=1.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
